Revert "VCST-2565: Edit configuration with files. Refactoring"
This reverts commit 55a560f3a43cd9e4c1b44e2f963f5105ef3fd7aa.
