--- src/VirtoCommerce.XCart.Core/Commands/Configuration/ChangeConfiguredLineItemCommand.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-using VirtoCommerce.CartModule.Core.Model;
-
-namespace VirtoCommerce.XCart.Core.Commands.Configuration;
-
-public class ChangeConfiguredLineItemCommand : ConfiguredLineItemCommand
-{
-    public IList<ConfigurationItem> ConfigurationItems { get; set; } = [];
-}
