--- src/VirtoCommerce.XCart.Core/Commands/Configuration/CreateConfiguredLineItemCommand.cs
@@ -1,5 +1,0 @@
-namespace VirtoCommerce.XCart.Core.Commands.Configuration;
-
-public class CreateConfiguredLineItemCommand : ConfiguredLineItemCommand
-{
-}
