+++ src/VirtoCommerce.XCart.Core/Commands/CreateConfiguredLineItemCommand.cs
@@ -2,9 +2,9 @@ using System.Collections.Generic;
 using VirtoCommerce.Xapi.Core.Infrastructure;
 using VirtoCommerce.XCart.Core.Models;
 
-namespace VirtoCommerce.XCart.Core.Commands.Configuration;
+namespace VirtoCommerce.XCart.Core.Commands;
 
-public abstract class ConfiguredLineItemCommand : ICommand<ExpConfigurationLineItem>, ICartProductContainerRequest
+public class CreateConfiguredLineItemCommand : ICommand<ExpConfigurationLineItem>, ICartProductContainerRequest
 {
     public string StoreId { get; set; }
