+++ src/VirtoCommerce.XCart.Data/Commands/AddCartItemCommandHandler.cs
@@ -8,7 +8,6 @@ using VirtoCommerce.Platform.Core.Common;
 using VirtoCommerce.XCart.Core;
 using VirtoCommerce.XCart.Core.Commands;
 using VirtoCommerce.XCart.Core.Commands.BaseCommands;
-using VirtoCommerce.XCart.Core.Commands.Configuration;
 using VirtoCommerce.XCart.Core.Models;
 using VirtoCommerce.XCart.Core.Services;
