+++ src/VirtoCommerce.XCart.Data/Commands/CreateConfiguredLineItemBuilder.cs
@@ -5,10 +5,10 @@ using Microsoft.AspNetCore.Authorization;
 using VirtoCommerce.Xapi.Core.BaseQueries;
 using VirtoCommerce.Xapi.Core.Extensions;
 using VirtoCommerce.XCart.Core;
-using VirtoCommerce.XCart.Core.Commands.Configuration;
+using VirtoCommerce.XCart.Core.Commands;
 using VirtoCommerce.XCart.Core.Schemas;
 
-namespace VirtoCommerce.XCart.Data.Commands.Configuration;
+namespace VirtoCommerce.XCart.Data.Commands;
 
 public class CreateConfiguredLineItemBuilder : CommandBuilder<CreateConfiguredLineItemCommand, ExpConfigurationLineItem, InputCreateConfiguredLineItemCommand, ConfigurationLineItemType>
 {
